+++ example/lib/main.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:native_admob_flutter/native_admob_flutter.dart';
 
+import 'screens/native_ads.dart';
+import 'screens/banner_ads.dart';
+
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   MobileAds.initialize();
@@ -17,6 +20,7 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
+      debugShowCheckedModeBanner: false,
       home: MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
@@ -34,142 +38,21 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.grey,
-      body: ListView(
-        children: [
-          // Create a native ad with the default style
-          NativeAd(
-            height: 110,
-            buildLayout: adBannerLayoutBuilder,
-            loading: Text('loading'),
-            error: Text('error'),
-            button: AdButtonView(
-              borderRadius: AdBorderRadius.vertical(bottom: 10),
-              margin: EdgeInsets.only(left: 6, right: 6, bottom: 6),
-              gradient:
-                  AdLinearGradient(colors: [Colors.yellow[300], Colors.amber]),
-            ),
+    return DefaultTabController(
+      length: 2,
+      child: Scaffold(
+        backgroundColor: Colors.grey,
+        appBar: AppBar(
+          title: Text('Ads demo'),
+          centerTitle: true,
+          bottom: TabBar(
+            tabs: [Tab(text: 'Native Ads'), Tab(text: 'Banner Ads')],
           ),
-          NativeAd(
-            height: 100,
-            buildLayout: secondBuilder,
-            loading: Text('loading'),
-            error: Text('error'),
-            icon: AdImageView(size: 80),
-            headline: AdTextView(
-              style: TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-              ),
-              maxLines: 1,
-            ),
-            media: AdMediaView(height: 80, width: 120),
-          ),
-          NativeAd(
-            height: 300,
-            buildLayout: fullBuilder,
-            loading: Text('loading'),
-            error: Text('error'),
-            icon: AdImageView(size: 40),
-            headline: AdTextView(
-              style: TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-              ),
-              maxLines: 1,
-            ),
-            media: AdMediaView(
-              height: 180,
-              width: MATCH_PARENT,
-            ),
-            attribution: AdTextView(
-              width: WRAP_CONTENT,
-              height: WRAP_CONTENT,
-              padding: EdgeInsets.symmetric(horizontal: 2, vertical: 0),
-              margin: EdgeInsets.only(right: 4),
-              maxLines: 1,
-              borderRadius: AdBorderRadius.all(10),
-              text: 'Anúncio',
-              border: BorderSide(color: Colors.green, width: 1),
-              style:
-                  TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
-            ),
-          ),
-          BannerAd(),
-        ],
+        ),
+        body: TabBarView(
+          children: [NativeAds(), BannerAds()],
+        ),
       ),
     );
   }
 }
-
-AdLayoutBuilder get fullBuilder => (ratingBar, media, icon, headline,
-        advertiser, body, price, store, attribuition, button) {
-      return AdLinearLayout(
-        padding: EdgeInsets.all(10),
-        // The first linear layout width needs to be extended to the
-        // parents height, otherwise the children won't fit good
-        width: MATCH_PARENT,
-        gradient: AdLinearGradient(
-          colors: [Colors.indigo[300], Colors.indigo[700]],
-          orientation: AdGradientOrientation.tl_br,
-        ),
-        children: [
-          media,
-          AdLinearLayout(
-            children: [
-              icon,
-              AdLinearLayout(children: [
-                headline,
-                AdLinearLayout(
-                  children: [attribuition, advertiser],
-                  orientation: HORIZONTAL,
-                  width: MATCH_PARENT,
-                ),
-              ], margin: EdgeInsets.only(left: 4)),
-            ],
-            width: WRAP_CONTENT,
-            orientation: HORIZONTAL,
-            margin: EdgeInsets.only(top: 6),
-          ),
-          AdLinearLayout(
-            children: [button],
-            orientation: HORIZONTAL,
-          ),
-        ],
-      );
-    };
-
-AdLayoutBuilder get secondBuilder => (ratingBar, media, icon, headline,
-        advertiser, body, price, store, attribution, button) {
-      return AdLinearLayout(
-        padding: EdgeInsets.all(10),
-        // The first linear layout width needs to be extended to the
-        // parents height, otherwise the children won't fit good
-        width: MATCH_PARENT,
-        orientation: HORIZONTAL,
-        gradient: AdRadialGradient(
-          colors: [Colors.blue[300], Colors.blue[900]],
-          center: Alignment(0.5, 0.5),
-          radius: 1000,
-        ),
-        children: [
-          icon,
-          AdLinearLayout(
-            children: [
-              headline,
-              AdLinearLayout(
-                children: [attribution, advertiser, ratingBar],
-                orientation: HORIZONTAL,
-                width: WRAP_CONTENT,
-                height: 20,
-              ),
-              button,
-            ],
-            margin: EdgeInsets.symmetric(horizontal: 4),
-          ),
-        ],
-      );
-    };